requirements-desktop.txt
-r requirements.txt
pywebview>=5,<6
pyinstaller>=6,<7
